Add transaction num to categories on manage data
Changes:
- Add badge to each category on the manage data page.
- Update all dependencies to latest version.
- Set descending date sort as the default.

Closes #97
