+++ package.json
@@ -1,40 +1,40 @@
 {
   "name": "off-the-books",
-  "version": "1.9.2",
+  "version": "1.10.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.4",
-    "@fortawesome/free-solid-svg-icons": "^5.3.1",
-    "@fortawesome/react-fontawesome": "^0.1.2",
+    "@fortawesome/fontawesome-svg-core": "^1.2.17",
+    "@fortawesome/free-solid-svg-icons": "^5.8.1",
+    "@fortawesome/react-fontawesome": "^0.1.4",
     "axios": "^0.18.0",
     "bayes": "^0.0.7",
-    "bootstrap": "^4.1.1",
-    "d3-array": "^1.2.1",
-    "d3-collection": "^1.0.4",
+    "bootstrap": "^4.3.1",
+    "d3-array": "^2.0.3",
+    "d3-collection": "^1.0.7",
     "d3-scale-chromatic": "^1.3.3",
     "downloadjs": "^1.4.7",
     "http-proxy-middleware": "^0.19.0",
-    "immutability-helper": "^2.6.6",
+    "immutability-helper": "^3.0.0",
     "jschardet": "^1.6.0",
     "lodash": "^4.17.10",
     "lru-cache": "^4.1.3",
-    "moment": "^2.22.2",
-    "netlify-lambda": "^0.4.0",
-    "papaparse": "^4.3.7",
-    "prop-types": "^15.6.1",
-    "react": "^16.4.2",
-    "react-dates": "^17.1.1",
-    "react-dom": "^16.4.2",
+    "moment": "^2.24.0",
+    "netlify-lambda": "^1.4.3",
+    "papaparse": "^4.6.3",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dates": "^20.1.0",
+    "react-dom": "^16.8.6",
     "react-loadable": "^5.5.0",
-    "react-redux": "^5.0.7",
+    "react-redux": "^6.0.1",
     "react-router-dom": "^4.2.2",
-    "react-scripts": "2.0.5",
-    "react-select": "^2.0.0",
-    "react-tooltip": "^3.5.1",
-    "recharts": "^1.1.0",
-    "redux": "^4.0.0",
+    "react-scripts": "2.1.8",
+    "react-select": "^2.4.2",
+    "react-tooltip": "^3.10.0",
+    "recharts": "^1.5.0",
+    "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
-    "redux-modal": "^2.0.3",
+    "redux-modal": "^3.0.1",
     "redux-persist": "^5.9.1",
     "redux-search": "^2.5.0",
     "redux-thunk": "^2.2.0",
@@ -51,9 +51,9 @@
     "analyze": "yarn build --stats && webpack-bundle-analyzer build/bundle-stats.json"
   },
   "devDependencies": {
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "jest-fetch-mock": "^1.6.6",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.11.2",
+    "jest-fetch-mock": "^2.1.1",
     "moxios": "^0.4.0",
     "redux-mock-store": "^1.5.1",
     "webpack-bundle-analyzer": "^3.0.3"
