+++ src/components/TransactionAdd.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import { MemoryRouter } from 'react-router-dom';
 import { mount } from 'enzyme';
@@ -38,7 +39,9 @@ describe('Transaction Add', () => {
 
     const container = mount(
       <MemoryRouter>
-        <TransactionAdd store={store} />
+        <Provider store={store}>
+          <TransactionAdd />
+        </Provider>
       </MemoryRouter>
     );
 
@@ -79,7 +82,9 @@ describe('Transaction Add', () => {
 
       const container = mount(
         <MemoryRouter>
-          <TransactionAdd store={store} />
+          <Provider store={store}>
+            <TransactionAdd />
+          </Provider>
         </MemoryRouter>
       );
 
@@ -120,7 +125,9 @@ describe('Transaction Add', () => {
 
       const container = mount(
         <MemoryRouter>
-          <TransactionAdd store={store} />
+          <Provider store={store}>
+            <TransactionAdd />
+          </Provider>
         </MemoryRouter>
       );
 
@@ -161,7 +168,9 @@ describe('Transaction Add', () => {
 
       const container = mount(
         <MemoryRouter>
-          <TransactionAdd store={store} />
+          <Provider store={store}>
+            <TransactionAdd />
+          </Provider>
         </MemoryRouter>
       );
 
@@ -199,7 +208,9 @@ describe('Transaction Add', () => {
 
       const container = mount(
         <MemoryRouter>
-          <TransactionAdd store={store} />
+          <Provider store={store}>
+            <TransactionAdd />
+          </Provider>
         </MemoryRouter>
       );
